refactor: migrated RasterAlgorithms to component packages under titiler group
