+++ packages/svelte-undp-components/src/lib/interfaces/RasterAlgorithm.ts
