+++ sites/geohub/src/components/util/stac/StacApiExplorer.svelte
@@ -16,7 +16,6 @@
 		DatasetFeature,
 		Layer,
 		LayerCreationInfo,
-		RasterAlgorithm,
 		Stac,
 		StacItemFeatureCollection,
 		StacProduct,
@@ -27,6 +26,7 @@
 		DatePicker,
 		FieldControl,
 		Notification,
+		type RasterAlgorithm,
 		type RasterTileMetadata,
 		SegmentButtons,
 		ShowDetails,
