+++ sites/geohub/src/components/util/stac/StacCatalogTools.svelte
@@ -6,8 +6,12 @@
 
 <script lang="ts">
 	import RasterAlgorithmExplorer from '$components/maplibre/raster/RasterAlgorithmExplorer.svelte';
-	import type { DatasetFeature, RasterAlgorithm, StacCollection, Tag } from '$lib/types';
-	import { Breadcrumbs, type BreadcrumbPage } from '@undp-data/svelte-undp-components';
+	import type { DatasetFeature, StacCollection, Tag } from '$lib/types';
+	import {
+		Breadcrumbs,
+		type BreadcrumbPage,
+		type RasterAlgorithm
+	} from '@undp-data/svelte-undp-components';
 	import { createEventDispatcher } from 'svelte';
 	import StacCatalogTool from './StacCatalogTool.svelte';
