+++ sites/geohub/src/routes/(app)/tools/+page.server.ts
@@ -3,7 +3,7 @@ import type { DatasetFeatureCollection, Tag } from '$lib/types';
 import type { PageServerLoad } from './$types';
 import { error } from '@sveltejs/kit';
 import { env } from '$env/dynamic/private';
-import { type RasterAlgorithm } from '$lib/types';
+import { type RasterAlgorithm } from '@undp-data/svelte-undp-components';
 
 export const load: PageServerLoad = async ({ fetch }) => {
 	const res = await fetch(`/api/tags?key=${ALGORITHM_TAG_KEY}`);
